translator: hoist divisions to correct Rust truncated div/mod
Cora's div/mod are Euclidean but Rust's truncate toward zero, so they
diverge for negative dividends. Hoist each division to its own program
point that binds a fresh variable to the sign-corrected value, instead of
substituting the correction inline (which explodes at nested divisions).

- width literal-only divisions from their context type, else i32
- while: re-run the hoisted condition each iteration via the back-edge
- name hoist vars $div ($ is legal in Cora, illegal in Rust, so it can't
  clash with a source variable); verified Cora parses and proves them
- regenerate division goldens; add while_div

+++ src/test/resources/benchmarks/endless_loop.rs
@@ -1,6 +1,6 @@
 // cora: MAYBE
-fn endless_loop() -> i32 {
-    let mut x: i32 = 0;
+fn endless_loop() -> i16 {
+    let mut x: i16 = 0;
     loop {
         x = x + 1;
     }
